pyproject.toml
[tool.poetry]
name = "nimble-test-flask"
version = "0.1.0"
description = ""
authors = ["Vaipik <n.khylko@gmail.com>"]
readme = "README.md"
packages = [{include = "nimble_test_flask"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-restful = "^0.3.9"
flask-sqlalchemy = "^3.0.2"
flask-migrate = "^4.0.1"
python-dotenv = "^0.21.0"
pyyaml = "^6.0"
uuid = "^1.30"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
